Add backup test extension with creating access method.
Add a extension test_backup as the test extension for WHPG-backup. It will create an access method.
Install the extension in Makefile
Add integration test case for GetAccessMethods() and end to end test case for restoring the extension with access method.
The extension can trigger the same error as pgvector, if we rollback PR #6

20250712:14:58:41 gprestore:lhl:rocky9:348696-[CRITICAL]:-ERROR: access method "backup_test_am" already exists (SQLSTATE 42710)